tests/requirements.txt
Flask-Testing==0.4.2
nose==1.3.7
